adjust classes to pass the testing (still not working)

+++ src/main/java/de/fhws/fiw/fds/suttondemo/client/rest/DemoRestClient.java
@@ -1,15 +1,16 @@
 package de.fhws.fiw.fds.suttondemo.client.rest;
 
 import de.fhws.fiw.fds.sutton.client.rest2.AbstractRestClient;
+import de.fhws.fiw.fds.sutton.client.utils.Link;
 import de.fhws.fiw.fds.suttondemo.client.models.ModuleClientModel;
 import de.fhws.fiw.fds.suttondemo.client.models.PartnerUniversityClientModel;
 import de.fhws.fiw.fds.suttondemo.client.web.ModuleWebClient;
 import de.fhws.fiw.fds.suttondemo.client.web.PartnerUniversityWebClient;
+import de.fhws.fiw.fds.suttondemo.server.api.states.modules.ModuleUri;
+import de.fhws.fiw.fds.suttondemo.server.api.states.partneruniversities.PartnerUniversityUri;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.List;
+import java.util.*;
 
 public class DemoRestClient extends AbstractRestClient {
     private static final String BASE_URL = "http://localhost:8080/demo/api";
@@ -45,12 +46,15 @@ public class DemoRestClient extends AbstractRestClient {
 
     public void start() throws IOException {
         processResponse(this.partnerUniversityClient.getDispatcher(BASE_URL), (response) -> {
-            System.out.println("Dispatcher started, available links: " + isLinkAvailable(GET_ALL_PARTNER_UNIVERSITIES) + ", " + isLinkAvailable(CREATE_PARTNER_UNIVERSITY) + ", " + isLinkAvailable(GET_ALL_MODULES) + ", " + isLinkAvailable(CREATE_MODULE));
+            System.out.println("Dispatcher started, available links: " + availableLinks());
         });
     }
 
     public boolean isCreatePartnerUniversityAllowed() {
-        return isLinkAvailable(CREATE_PARTNER_UNIVERSITY);
+        boolean allowed = isLinkAvailable(CREATE_PARTNER_UNIVERSITY);
+        System.out.println("isCreatePartnerUniversityAllowed: " + allowed);
+        System.out.println("Available links: " + availableLinks());
+        return allowed;
     }
 
     public void createPartnerUniversity(PartnerUniversityClientModel partnerUniversity) throws IOException {
@@ -65,7 +69,9 @@ public class DemoRestClient extends AbstractRestClient {
     }
 
     public boolean isGetAllPartnerUniversitiesAllowed() {
-        return isLinkAvailable(GET_ALL_PARTNER_UNIVERSITIES);
+        boolean allowed = isLinkAvailable(GET_ALL_PARTNER_UNIVERSITIES);
+        System.out.println("isGetAllPartnerUniversitiesAllowed: " + allowed);
+        return allowed;
     }
 
     public void getAllPartnerUniversities() throws IOException {
@@ -80,16 +86,16 @@ public class DemoRestClient extends AbstractRestClient {
     }
 
     public boolean isGetSinglePartnerUniversityAllowed() {
-        return !this.currentPartnerUniversityData.isEmpty() || isLocationHeaderAvailable();
+        boolean allowed = isLinkAvailable(PartnerUniversityUri.REL_PATH_ID);
+        System.out.println("isGetSinglePartnerUniversityAllowed: " + allowed);
+        return allowed;
     }
 
-    public void getSinglePartnerUniversity() throws IOException {
-        if (isLocationHeaderAvailable()) {
-            getSinglePartnerUniversity(getLocationHeaderURL());
-        } else if (!this.currentPartnerUniversityData.isEmpty()) {
-            getSinglePartnerUniversity(this.cursorPartnerUniversityData);
+    public void getSinglePartnerUniversity(long createdId) throws IOException {
+        if (isGetSinglePartnerUniversityAllowed()) {
+            getSinglePartnerUniversity(getPartnerUniversityUrl(createdId));
         } else {
-            throw new IllegalStateException();
+            throw new IllegalStateException("Get Single Partner University not allowed");
         }
     }
 
@@ -97,10 +103,14 @@ public class DemoRestClient extends AbstractRestClient {
         getSinglePartnerUniversity(this.currentPartnerUniversityData.get(index).getSelfLink().getUrl());
     }
 
-    private void getSinglePartnerUniversity(String url) throws IOException {
+    public void getSinglePartnerUniversity(String url) throws IOException {
         processResponse(this.partnerUniversityClient.getSinglePartnerUniversity(url), (response) -> {
-            this.currentPartnerUniversityData = new LinkedList<>(response.getResponseData());
-            this.cursorPartnerUniversityData = 0;
+            Collection<PartnerUniversityClientModel> responseData = response.getResponseData();
+            if (!responseData.isEmpty()) {
+                PartnerUniversityClientModel partnerUniversity = responseData.iterator().next();
+                this.currentPartnerUniversityData = List.of(partnerUniversity);
+                this.cursorPartnerUniversityData = 0;
+            }
         });
     }
 
@@ -119,25 +129,61 @@ public class DemoRestClient extends AbstractRestClient {
         }
     }
 
-    public void updatePartnerUniversity(String url, PartnerUniversityClientModel university) throws IOException {
-        processResponse(this.partnerUniversityClient.putPartnerUniversity(url, university), (response) -> {
-            // Update current partner university data if needed
-        });
+    public boolean isUpdatePartnerUniversityAllowed() {
+        boolean allowed = isLinkAvailable(UPDATE_PARTNER_UNIVERSITY);
+        System.out.println("isUpdatePartnerUniversityAllowed: " + allowed);
+        return allowed;
     }
 
+    public void updatePartnerUniversity(String url, PartnerUniversityClientModel university) throws IOException {
+        if (isUpdatePartnerUniversityAllowed()) {
+            processResponse(this.partnerUniversityClient.putPartnerUniversity(url, university), (response) -> {
+                try {
+                    getAllPartnerUniversities(); // Refresh the data
+                } catch (IOException e) {
+                    e.printStackTrace(); // Handle the exception appropriately
+                }
+            });
+        } else {
+            throw new IllegalStateException("Update Partner University not allowed");
+        }
+    }
+
+    public boolean isDeletePartnerUniversityAllowed() throws IOException {
+        String availableLinks = getAvailableLinks();
+        System.out.println("Checking if deletePartnerUniversity is allowed...");
+        System.out.println("Available links: " + availableLinks);
+        boolean isAllowed = availableLinks.contains("deletePartnerUniversity ->");
+        System.out.println("isDeletePartnerUniversityAllowed: " + isAllowed);
+        return isAllowed;
+    }
+
+
+
     public void deletePartnerUniversity(String url) throws IOException {
-        processResponse(this.partnerUniversityClient.deletePartnerUniversity(url), (response) -> {
-            // Update current partner university data if needed
-        });
+        if (isDeletePartnerUniversityAllowed()) {
+            processResponse(this.partnerUniversityClient.deletePartnerUniversity(url), (response) -> {
+                try {
+                    getAllPartnerUniversities(); // Refresh the data
+                } catch (IOException e) {
+                    e.printStackTrace(); // Handle the exception appropriately
+                }
+            });
+        } else {
+            throw new IllegalStateException("Delete Partner University not allowed");
+        }
     }
 
     public boolean isCreateModuleAllowed() {
-        return isLinkAvailable(CREATE_MODULE);
+        boolean allowed = isLinkAvailable(CREATE_MODULE);
+        System.out.println("isCreateModuleAllowed: " + allowed);
+        return allowed;
     }
 
-    public void createModule(ModuleClientModel module) throws IOException {
+    public void createModule(ModuleClientModel module, long partnerUniversityId) throws IOException {
         if (isCreateModuleAllowed()) {
-            processResponse(this.moduleClient.postNewModule(getUrl(CREATE_MODULE), module), (response) -> {
+            String moduleUrl = getModuleCreationUrl(partnerUniversityId);
+            processResponse(this.moduleClient.postNewModule(moduleUrl, module), (response) -> {
                 this.currentModuleData = Collections.emptyList();
                 this.cursorModuleData = 0;
             });
@@ -147,12 +193,15 @@ public class DemoRestClient extends AbstractRestClient {
     }
 
     public boolean isGetAllModulesAllowed() {
-        return isLinkAvailable(GET_ALL_MODULES);
+        boolean allowed = isLinkAvailable(GET_ALL_MODULES);
+        System.out.println("isGetAllModulesAllowed: " + allowed);
+        return allowed;
     }
 
-    public void getAllModules() throws IOException {
+    public void getAllModules(long partnerUniversityId) throws IOException {
         if (isGetAllModulesAllowed()) {
-            processResponse(this.moduleClient.getCollectionOfModules(getUrl(GET_ALL_MODULES)), (response) -> {
+            String moduleUrl = getAllModulesUrl(partnerUniversityId);
+            processResponse(this.moduleClient.getCollectionOfModules(moduleUrl), (response) -> {
                 this.currentModuleData = new LinkedList<>(response.getResponseData());
                 this.cursorModuleData = 0;
             });
@@ -162,16 +211,16 @@ public class DemoRestClient extends AbstractRestClient {
     }
 
     public boolean isGetSingleModuleAllowed() {
-        return !this.currentModuleData.isEmpty() || isLocationHeaderAvailable();
+        boolean allowed = isLinkAvailable(ModuleUri.REL_PATH_ID);
+        System.out.println("isGetSingleModuleAllowed: " + allowed);
+        return allowed;
     }
 
-    public void getSingleModule() throws IOException {
-        if (isLocationHeaderAvailable()) {
-            getSingleModule(getLocationHeaderURL());
-        } else if (!this.currentModuleData.isEmpty()) {
-            getSingleModule(this.cursorModuleData);
+    public void getSingleModule(long universityId, long moduleId) throws IOException {
+        if (isGetSingleModuleAllowed()) {
+            getSingleModule(getModuleUrl(universityId, moduleId));
         } else {
-            throw new IllegalStateException();
+            throw new IllegalStateException("Get Single Module not allowed");
         }
     }
 
@@ -181,21 +230,54 @@ public class DemoRestClient extends AbstractRestClient {
 
     private void getSingleModule(String url) throws IOException {
         processResponse(this.moduleClient.getSingleModule(url), (response) -> {
-            this.currentModuleData = new LinkedList<>(response.getResponseData());
-            this.cursorModuleData = 0;
+            Collection<ModuleClientModel> responseData = response.getResponseData();
+            if (!responseData.isEmpty()) {
+                ModuleClientModel module = responseData.iterator().next();
+                this.currentModuleData = List.of(module);
+                this.cursorModuleData = 0;
+            }
         });
+    }
+
+    public boolean isUpdateModuleAllowed() {
+        boolean allowed = isLinkAvailable(UPDATE_MODULE);
+        System.out.println("isUpdateModuleAllowed: " + allowed);
+        return allowed;
     }
 
     public void updateModule(String url, ModuleClientModel module) throws IOException {
-        processResponse(this.moduleClient.putModule(url, module), (response) -> {
-            // Update current module data if needed
-        });
+        if (isUpdateModuleAllowed()) {
+            processResponse(this.moduleClient.putModule(url, module), (response) -> {
+                try {
+                    getAllModules(module.getPartnerUniversityId()); // Refresh the data
+                } catch (IOException e) {
+                    e.printStackTrace(); // Handle the exception appropriately
+                }
+            });
+        } else {
+            throw new IllegalStateException("Update Module not allowed");
+        }
+    }
+
+    public boolean isDeleteModuleAllowed() {
+        boolean allowed = isLinkAvailable(DELETE_MODULE);
+        System.out.println("isDeleteModuleAllowed: " + allowed);
+        return allowed;
     }
 
     public void deleteModule(String url) throws IOException {
-        processResponse(this.moduleClient.deleteModule(url), (response) -> {
-            // Update current module data if needed
-        });
+        if (isDeleteModuleAllowed()) {
+            processResponse(this.moduleClient.deleteModule(url), (response) -> {
+                long partnerUniversityId = extractPartnerUniversityIdFromModuleUrl(url);
+                try {
+                    getAllModules(partnerUniversityId); // Refresh the data
+                } catch (IOException e) {
+                    e.printStackTrace(); // Handle the exception appropriately
+                }
+            });
+        } else {
+            throw new IllegalStateException("Delete Module not allowed");
+        }
     }
 
     public List<ModuleClientModel> moduleData() {
@@ -213,11 +295,40 @@ public class DemoRestClient extends AbstractRestClient {
         }
     }
 
-    public String getModuleUrl(long moduleId) {
-        return String.format("%s/modules/%d", BASE_URL, moduleId);
+    public String getModuleCreationUrl(long partnerUniversityId) {
+        return String.format("%s/partneruniversities/%d/modules", BASE_URL, partnerUniversityId);
+    }
+
+    public String getAllModulesUrl(long partnerUniversityId) {
+        return String.format("%s/partneruniversities/%d/modules", BASE_URL, partnerUniversityId);
+    }
+
+    public String getModuleUrl(long universityId, long moduleId) {
+        return String.format("%s/partneruniversities/%d/modules/%d", BASE_URL, universityId, moduleId);
     }
 
     public String getPartnerUniversityUrl(long universityId) {
         return String.format("%s/partneruniversities/%d", BASE_URL, universityId);
+    }
+
+    private long extractPartnerUniversityIdFromModuleUrl(String url) {
+        String[] parts = url.split("/");
+        return Long.parseLong(parts[parts.length - 3]);
+    }
+
+    public String availableLinks() {
+        StringBuilder builder = new StringBuilder();
+        for (Map.Entry<String, Link> entry : getPossibleNextStates().entrySet()) {
+            builder.append(entry.getKey()).append(" -> ").append(entry.getValue().getUrl()).append("\n");
+        }
+        return builder.toString();
+    }
+
+    public String getAvailableLinks() {
+        StringBuilder builder = new StringBuilder();
+        for (Map.Entry<String, Link> entry : getPossibleNextStates().entrySet()) {
+            builder.append(entry.getKey()).append(" -> ").append(entry.getValue().getUrl()).append("\n");
+        }
+        return builder.toString();
     }
 }
